Split Admin CLI into modules per technology
* Introduce `polairs-admin-base` for shared commands

* Move NoSQL commands to `polaris-persistence-nosql-admin`

* Make NoSQL commands optional with build property: `ExcludeNoSqlPersistence` as proposed in #3867

* Move JDBC-specific dependencies for admin commands to `polaris-relational-jdbc-admin`

* CDI-based Sub-command discovery is handled by `AdminCliConfiguration`

* Add dedicated admin test modules per persistence technology (JDBC / NoSQL)

* Move default quarkus properties to `runtime-admin-defaults` to allow downstream projects to _not_ include them.

* Downstream builds are advised to depend on individual command modules
  (noted above) and have local `application.properties`, but not depend
  on `runtime-admin`, which serves as the Admin CLI distribution assembly
  module for Apache Polaris releases.

Closes #3867

This change also addresses most of concerns from #3855.
