+++ gradle/projects.main.properties
@@ -30,9 +30,13 @@ polaris-server=runtime/server
 polaris-distribution=runtime/distribution
 polaris-runtime-spark-tests=runtime/spark-tests
 polaris-admin=runtime/admin
+polaris-admin-base=runtime/admin-base
+polaris-admin-defaults=runtime/admin-defaults
 polaris-runtime-common=runtime/common
 polaris-runtime-test-common=runtime/test-common
 polaris-relational-jdbc=persistence/relational-jdbc
+polaris-relational-jdbc-admin=persistence/relational-jdbc-admin/cmd
+polaris-relational-jdbc-admin-test=persistence/relational-jdbc-admin/test
 polaris-tests=integration-tests
 aggregated-license-report=aggregated-license-report
 polaris-immutables=tools/immutables
@@ -97,3 +101,5 @@ polaris-persistence-nosql-maintenance-spi=persistence/nosql/persistence/maintena
 # persistence / database specific implementations
 polaris-persistence-nosql-inmemory=persistence/nosql/persistence/db/inmemory
 polaris-persistence-nosql-mongodb=persistence/nosql/persistence/db/mongodb
+polaris-persistence-nosql-admin=persistence/nosql/admin/cmd
+polaris-persistence-nosql-admin-test=persistence/nosql/admin/test
