+++ persistence/nosql/admin/cmd/src/main/java/org/apache/polaris/admintool/nosql/Placeholder.java
@@ -1,0 +1,31 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.polaris.admintool.nosql;
+
+import jakarta.enterprise.context.ApplicationScoped;
+
+/**
+ * Something with a "bean-producing" annotation is required by Quarkus to "notice" this module, yet
+ * Quarkus will find CLI commands only if they are `@Default`, so `@ApplicationScoped` cannot be
+ * applied to them... hence the workaround with this placeholder just to hold the
+ * `@ApplicationScoped` annotation.
+ */
+@ApplicationScoped
+public class Placeholder {}
