+++ persistence/nosql/admin/test/src/test/java/org/apache/polaris/admintool/nosql/NoSqlInMemoryProfile.java
