+++ runtime/admin-base/src/main/java/org/apache/polaris/admintool/BaseCommand.java
