+++ runtime/admin-base/src/testFixtures/resources/org/apache/polaris/admintool/credentials.json
