--- runtime/admin/src/main/resources/application.properties
@@ -1,88 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-# ---- Build-time configuration ----
-# Build-time properties take effect when Quarkus processes java classes
-# and creates Polaris jars. These properties cannot be overridden in runtime.
-# Cf. https://quarkus.io/guides/all-config
-
-quarkus.application.name=Apache Polaris Admin Tool
-quarkus.banner.enabled=false
-
-quarkus.container-image.build=false
-quarkus.container-image.push=false
-quarkus.container-image.registry=docker.io
-quarkus.container-image.group=apache
-quarkus.container-image.name=polaris-admin-tool
-quarkus.container-image.additional-tags=latest
-quarkus.datasource.db-kind=postgresql
-# if set to true it will try to start localstack at build and run time for the local environment
-# https://docs.quarkiverse.io/quarkus-amazon-services/dev/amazon-rds.html#_configuration_reference for more details
-quarkus.rds.devservices.enabled=false
-quarkus.rds.sync-client.type=apache
-quarkus.mongodb.devservices.enabled=false
-
-# ---- Runtime Configuration ----
-# Below are default values for properties that can be changed in runtime.
-
-# Available types:
-# - in-memory - InMemoryPolarisMetaStoreManagerFactory
-# - in-memory-atomic - InMemoryAtomicOperationMetaStoreManagerFactory
-# - relational-jdbc - JdbcMetaStoreManagerFactory
-# - nosql - NoSQL persistence backend, define the backend type via 'polaris.persistence.nosql.backend'
-polaris.persistence.type=relational-jdbc
-# Database backend for 'nosql' persistence-type
-# Available backends:
-# - InMemory - for testing purposes
-# - MongoDb - configure the via the Quarkus extension
-#     Configure the necessary MongoDB properties starting with 'quarkus.mongodb.' in this file.
-#     See https://quarkus.io/guides/mongodb#configuration-reference for details about these configurations.
-#polaris.persistence.nosql.backend=InMemory
-
-## MongoDB version store specific configuration
-#quarkus.mongodb.database=polaris
-#quarkus.mongodb.metrics.enabled=true
-#quarkus.mongodb.connection-string=mongodb://localhost:27017
-
-#quarkus.datasource.active=false
-#quarkus.datasource.url=
-#quarkus.datasource.username=
-#quarkus.datasource.password=
-
-quarkus.datasource.devservices.enabled=false
-
-quarkus.arc.ignored-split-packages=\
-  org.apache.polaris.admintool.config,\
-  org.apache.polaris.admintool
-
-## Quarkus required setting for third party indexing
-# fixed at build-time
-quarkus.index-dependency.avro.group-id=org.apache.avro
-quarkus.index-dependency.avro.artifact-id=avro
-quarkus.index-dependency.guava.group-id=com.google.guava
-quarkus.index-dependency.guava.artifact-id=guava
-quarkus.index-dependency.protobuf.group-id=com.google.protobuf
-quarkus.index-dependency.protobuf.artifact-id=protobuf-java
-
-## Logging Configuration
-# Set root logging level to WARN to avoid mostly irrelevant Quarkus log messages.
-# The Admin tool communicates with the user via STDOUT/STDERR
-quarkus.log.level=WARN
-# Prevent the 'The Agroal dependency is present but no JDBC datasources have been defined.' build-time warning.
-quarkus.log.category."io.quarkus.agroal.deployment".level=ERROR
